[FIX+IMP] l10n_es_facturae: avoid error with mode == 'modify'  + don't write several times
When doing the refund, if you select third option (modify - create refund, reconcile and
create a new one), the returned action doesn't contain a domain, but a `res_id`, resulting
an error trying to interpret the domain that comes as a string (but better this than having
a write over full refunds).

We fix it injecting values in context and later rescuing them on refund values dictionary
preparation than doing the old way.

This also improves the performance, not having to write again to DB for adding these values.