[FIX] l10n_es_facturae: The "global" tax calculation rounding method is always used to compute the tax amounts even if the company has the tax rounding method set to "per line"
This commit fixes an issue in the l10n_es_facturae module where the "global" tax calculation rounding method was always used, regardless of the company's tax rounding method setting. Before the fix, even if the company had the tax rounding method set to "per line", the module incorrectly applied the "global" method when generating the XML file.

The changes implemented in this commit ensure that the correct tax rounding method, as defined by the company's settings, is used for tax calculations.