[FIX] l10n_es_facturae: don't copy facturae flag
The origin is an incompatibility detected with partner_vat_unique. That
module sets the res.partner vat field to copy=False (so it's unique). In
that case, when we try to copy a partner, if it has the facturae flag
set, the it will fail as it won't have a VAT number.

TT51626